# perf: Reduce unneeded hashing when loading provinces