--- a/src/parse/localization.rs
+++ b/src/parse/localization.rs
@@ -438,28 +438,33 @@ impl<'a> ValueParser<'a> {
         }
     }
 
-    fn peek(&mut self) -> Option<char> {
-        let p = self.content_iters[self.content_idx].peek();
-        if p.is_none() {
-            if self.content_idx + 1 == self.content.len() {
-                None
-            } else {
-                self.content_idx += 1;
-                self.loc = self.content[self.content_idx].loc;
-                self.offset = 0;
-                self.peek()
-            }
+    fn maybe_advance_idx(&mut self) -> bool {
+        if self.content_idx + 1 == self.content.len() {
+            false
         } else {
-            p.copied()
+            self.content_idx += 1;
+            self.loc = self.content[self.content_idx].loc;
+            self.offset = 0;
+            true
+        }
+    }
+
+    fn peek_or_advance_idx(&mut self) -> Option<char> {
+        if let Some(p) = self.content_iters[self.content_idx].peek() {
+            Some(*p)
+        } else if self.maybe_advance_idx() {
+            self.peek_or_advance_idx()
+        } else {
+            None
         }
     }
 
     fn next_char(&mut self) {
-        // self.peek advances content_idx if needed
-        if self.peek().is_some() {
-            let c = self.content_iters[self.content_idx].next().unwrap();
+        if let Some(c) = self.content_iters[self.content_idx].next() {
             self.offset += c.len_utf8();
             self.loc.column += 1;
+        } else if self.maybe_advance_idx() {
+            self.next_char();
         }
     }
 
@@ -470,7 +475,7 @@ impl<'a> ValueParser<'a> {
     }
 
     fn skip_whitespace(&mut self) {
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             if c.is_whitespace() {
                 self.next_char();
             } else {
@@ -481,14 +486,14 @@ impl<'a> ValueParser<'a> {
 
     fn unexpected_char(&mut self, expected: &str, errorkey: ErrorKey) {
         // TODO: handle EOF better
-        let c = self.peek().unwrap_or(' ');
+        let c = self.peek_or_advance_idx().unwrap_or(' ');
         let msg = format!("Unexpected character `{c}`, {expected}");
         warn(errorkey).msg(msg).loc(self.loc).push();
     }
 
     fn get_key(&mut self) -> Token {
         let mut text = self.start_text();
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             if is_key_char(c) {
                 text.add_char(c);
                 self.next_char();
@@ -500,10 +505,10 @@ impl<'a> ValueParser<'a> {
     }
 
     fn parse_format(&mut self) -> Option<Token> {
-        (self.peek() == Some('|')).then(|| {
+        (self.peek_or_advance_idx() == Some('|')).then(|| {
             self.next_char(); // eat the |
             let mut text = self.start_text();
-            while let Some(c) = self.peek() {
+            while let Some(c) = self.peek_or_advance_idx() {
                 if c == '$' || c == ']' {
                     break;
                 }
@@ -520,12 +525,12 @@ impl<'a> ValueParser<'a> {
 
         loop {
             self.skip_whitespace();
-            if self.peek() == Some('\'') {
+            if self.peek_or_advance_idx() == Some('\'') {
                 self.next_char();
                 let loc = self.loc;
                 let mut parens: isize = 0;
                 let mut text = self.start_text();
-                while let Some(c) = self.peek() {
+                while let Some(c) = self.peek_or_advance_idx() {
                     match c {
                         '\'' => break,
                         ']' => {
@@ -548,24 +553,24 @@ impl<'a> ValueParser<'a> {
                     text.add_char(c);
                     self.next_char();
                 }
-                if self.peek() != Some('\'') {
+                if self.peek_or_advance_idx() != Some('\'') {
                     self.value.push(LocaValue::Error);
                     return Vec::new();
                 }
                 v.push(CodeArg::Literal(text.take_to_token()));
                 self.next_char();
-            } else if self.peek() == Some(')') {
+            } else if self.peek_or_advance_idx() == Some(')') {
                 // Empty () means no arguments
             } else {
                 v.push(CodeArg::Chain(self.parse_code_inner()));
             }
             self.skip_whitespace();
-            if self.peek() != Some(',') {
+            if self.peek_or_advance_idx() != Some(',') {
                 break;
             }
             self.next_char(); // Eat the comma
         }
-        if self.peek() == Some(')') {
+        if self.peek_or_advance_idx() == Some(')') {
             self.next_char();
         } else {
             self.unexpected_char("expected `)`", ErrorKey::Datafunctions);
@@ -576,12 +581,12 @@ impl<'a> ValueParser<'a> {
     fn parse_code_code(&mut self) -> Code {
         let mut text = self.start_text();
 
-        if Game::is_hoi4() && self.peek() == Some('?') {
+        if Game::is_hoi4() && self.peek_or_advance_idx() == Some('?') {
             text.add_char('?');
             self.next_char();
         }
 
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             if is_code_char(c) {
                 text.add_char(c);
                 self.next_char();
@@ -590,7 +595,7 @@ impl<'a> ValueParser<'a> {
             }
         }
         let name = text.take_to_token();
-        if self.peek() == Some('(') {
+        if self.peek_or_advance_idx() == Some('(') {
             Code { name, arguments: self.parse_code_args() }
         } else {
             Code { name, arguments: Vec::new() }
@@ -603,11 +608,11 @@ impl<'a> ValueParser<'a> {
             v.push(self.parse_code_code());
             // Newlines followed by whitespace are allowed in code sequences,
             // but not random whitespace.
-            if matches!(self.peek(), Some('\r' | '\n')) {
+            if matches!(self.peek_or_advance_idx(), Some('\r' | '\n')) {
                 self.next_char();
                 self.skip_whitespace();
             }
-            if self.peek() != Some('.') {
+            if self.peek_or_advance_idx() != Some('.') {
                 break;
             }
             self.next_char(); // Eat the '.'
@@ -626,7 +631,7 @@ impl<'a> ValueParser<'a> {
         // The game engine doesn't mind if there are too many `)` before the `]`, so handle
         // that at "untidy" severity.
         let mut warned_extra_parens = false;
-        while self.peek() == Some(')') {
+        while self.peek_or_advance_idx() == Some(')') {
             if !warned_extra_parens {
                 let msg = "too many `)`";
                 untidy(ErrorKey::Datafunctions).msg(msg).loc(self.loc).push();
@@ -638,7 +643,7 @@ impl<'a> ValueParser<'a> {
 
         let format = self.parse_format();
 
-        if self.peek() == Some(']') {
+        if self.peek_or_advance_idx() == Some(']') {
             self.next_char();
             self.value.push(LocaValue::Code(chain, format));
         } else {
@@ -665,11 +670,11 @@ impl<'a> ValueParser<'a> {
     fn parse_markup(&mut self) {
         let loc = self.loc;
         self.next_char(); // skip the #
-        if self.peek() == Some('#') {
+        if self.peek_or_advance_idx() == Some('#') {
             // double # means a literal #
             self.next_char();
             self.value.push(LocaValue::Text(Token::from_static_str("#", loc)));
-        } else if self.peek() == Some('!') {
+        } else if self.peek_or_advance_idx() == Some('!') {
             self.next_char();
             self.value.push(LocaValue::MarkupEnd);
         } else {
@@ -685,7 +690,7 @@ impl<'a> ValueParser<'a> {
                 InValue(String, String, Loc, usize),
             }
             let mut state = State::InKey(String::new());
-            while let Some(c) = self.peek() {
+            while let Some(c) = self.peek_or_advance_idx() {
                 if c.is_whitespace() {
                     break;
                 }
@@ -771,7 +776,7 @@ impl<'a> ValueParser<'a> {
                     }
                 }
             }
-            if self.peek().is_none_or(char::is_whitespace) {
+            if self.peek_or_advance_idx().is_none_or(char::is_whitespace) {
                 self.next_char();
             } else {
                 let msg = "#markup should be followed by a space";
@@ -786,7 +791,7 @@ impl<'a> ValueParser<'a> {
 
         let mut old_value = take(&mut self.value);
 
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             if c == '[' {
                 self.parse_code();
             } else if is_key_char(c) {
@@ -830,7 +835,7 @@ impl<'a> ValueParser<'a> {
         self.next_char(); // eat the @
 
         let mut text = self.start_text();
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             if c.is_ascii_uppercase() {
                 text.add_char(c);
                 self.next_char();
@@ -850,7 +855,7 @@ impl<'a> ValueParser<'a> {
     fn parse_escape(&mut self) {
         let loc = self.loc;
         self.next_char(); // Skip the \
-        let s = match self.peek() {
+        let s = match self.peek_or_advance_idx() {
             Some('n') => '\n'.to_string(),
             Some(c) => c.to_string(),
             None => {
@@ -864,7 +869,7 @@ impl<'a> ValueParser<'a> {
 
     fn parse_text(&mut self) {
         let mut text = self.start_text();
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             match c {
                 '[' | '#' | '@' | '\\' => break,
                 _ => {
@@ -877,7 +882,7 @@ impl<'a> ValueParser<'a> {
     }
 
     pub fn parse_vec(mut self) -> Vec<LocaValue> {
-        while let Some(c) = self.peek() {
+        while let Some(c) = self.peek_or_advance_idx() {
             match c {
                 '[' => self.parse_code(),
                 '#' => self.parse_markup(),