@amtep 2025-07-27
A README.md with instructions for running the benchmarks would be nice :)

For me it was:
* create `internal_benches/ck3.toml` (needs example)
* create `internal_benches/vic3.toml` (needs example)
* run `cargo bench --features ck3 -p internal_benches`
* run `cargo bench --features vic3 -p internal_benches`